@@ -25,21 +25,23 @@ var customAPIWidgetTemplate = mustParseTemplate("custom-api.html", "widget-base.
 
 // Needs to be exported for the YAML unmarshaler to work
 type CustomAPIRequest struct {
-	URL           string               `yaml:"url"`
-	AllowInsecure bool                 `yaml:"allow-insecure"`
-	Headers       map[string]string    `yaml:"headers"`
-	Parameters    queryParametersField `yaml:"parameters"`
-	Method        string               `yaml:"method"`
-	BodyType      string               `yaml:"body-type"`
-	Body          any                  `yaml:"body"`
-	bodyReader    io.ReadSeeker        `yaml:"-"`
-	httpRequest   *http.Request        `yaml:"-"`
+	URL                string               `yaml:"url"`
+	AllowInsecure      bool                 `yaml:"allow-insecure"`
+	Headers            map[string]string    `yaml:"headers"`
+	Parameters         queryParametersField `yaml:"parameters"`
+	Method             string               `yaml:"method"`
+	BodyType           string               `yaml:"body-type"`
+	Body               any                  `yaml:"body"`
+	SkipJSONValidation bool                 `yaml:"skip-json-validation"`
+	bodyReader         io.ReadSeeker        `yaml:"-"`
+	httpRequest        *http.Request        `yaml:"-"`
 }
 
 type customAPIWidget struct {
 	widgetBase        `yaml:",inline"`
 	*CustomAPIRequest `yaml:",inline"`             // the primary request
 	Subrequests       map[string]*CustomAPIRequest `yaml:"subrequests"`
+	Options           customAPIOptions             `yaml:"options"`
 	Template          string                       `yaml:"template"`
 	Frameless         bool                         `yaml:"frameless"`
 	compiledTemplate  *template.Template           `yaml:"-"`
@@ -74,7 +76,9 @@ func (widget *customAPIWidget) initialize() error {
 }
 
 func (widget *customAPIWidget) update(ctx context.Context) {
-	compiledHTML, err := fetchAndParseCustomAPI(widget.CustomAPIRequest, widget.Subrequests, widget.compiledTemplate)
+	compiledHTML, err := fetchAndRenderCustomAPIRequest(
+		widget.CustomAPIRequest, widget.Subrequests, widget.Options, widget.compiledTemplate,
+	)
 	if !widget.canContinueUpdateAfterHandlingErr(err) {
 		return
 	}
@@ -86,9 +90,50 @@ func (widget *customAPIWidget) Render() template.HTML {
 	return widget.renderTemplate(widget, customAPIWidgetTemplate)
 }
 
+type customAPIOptions map[string]any
+
+func (o *customAPIOptions) StringOr(key, defaultValue string) string {
+	return customAPIGetOptionOrDefault(*o, key, defaultValue)
+}
+
+func (o *customAPIOptions) IntOr(key string, defaultValue int) int {
+	return customAPIGetOptionOrDefault(*o, key, defaultValue)
+}
+
+func (o *customAPIOptions) FloatOr(key string, defaultValue float64) float64 {
+	return customAPIGetOptionOrDefault(*o, key, defaultValue)
+}
+
+func (o *customAPIOptions) BoolOr(key string, defaultValue bool) bool {
+	return customAPIGetOptionOrDefault(*o, key, defaultValue)
+}
+
+func (o *customAPIOptions) JSON(key string) string {
+	value, exists := (*o)[key]
+	if !exists {
+		panic(fmt.Sprintf("key %q does not exist in options", key))
+	}
+
+	encoded, err := json.Marshal(value)
+	if err != nil {
+		panic(fmt.Sprintf("marshaling %s: %v", key, err))
+	}
+
+	return string(encoded)
+}
+
+func customAPIGetOptionOrDefault[T any](o customAPIOptions, key string, defaultValue T) T {
+	if value, exists := o[key]; exists {
+		if typedValue, ok := value.(T); ok {
+			return typedValue
+		}
+	}
+	return defaultValue
+}
+
 func (req *CustomAPIRequest) initialize() error {
-	if req.URL == "" {
-		return errors.New("URL is required")
+	if req == nil || req.URL == "" {
+		return nil
 	}
 
 	if req.Body != nil {
@@ -155,6 +200,18 @@ type customAPIResponseData struct {
 type customAPITemplateData struct {
 	*customAPIResponseData
 	subrequests map[string]*customAPIResponseData
+	Options     customAPIOptions
+}
+
+func (data *customAPITemplateData) JSONLines() []decoratedGJSONResult {
+	result := make([]decoratedGJSONResult, 0, 5)
+
+	gjson.ForEachLine(data.JSON.Raw, func(line gjson.Result) bool {
+		result = append(result, decoratedGJSONResult{line})
+		return true
+	})
+
+	return result
 }
 
 func (data *customAPITemplateData) Subrequest(key string) *customAPIResponseData {
@@ -171,7 +228,14 @@ func (data *customAPITemplateData) Subrequest(key string) *customAPIResponseData
 	return req
 }
 
-func fetchCustomAPIRequest(ctx context.Context, req *CustomAPIRequest) (*customAPIResponseData, error) {
+func fetchCustomAPIResponse(ctx context.Context, req *CustomAPIRequest) (*customAPIResponseData, error) {
+	if req == nil || req.URL == "" {
+		return &customAPIResponseData{
+			JSON:     decoratedGJSONResult{gjson.Result{}},
+			Response: &http.Response{},
+		}, nil
+	}
+
 	if req.bodyReader != nil {
 		req.bodyReader.Seek(0, io.SeekStart)
 	}
@@ -190,27 +254,31 @@ func fetchCustomAPIRequest(ctx context.Context, req *CustomAPIRequest) (*customA
 
 	body := strings.TrimSpace(string(bodyBytes))
 
-	if body != "" && !gjson.Valid(body) {
-		truncatedBody, isTruncated := limitStringLength(body, 100)
-		if isTruncated {
-			truncatedBody += "... <truncated>"
+	if !req.SkipJSONValidation && body != "" && !gjson.Valid(body) {
+		if 200 <= resp.StatusCode && resp.StatusCode < 300 {
+			truncatedBody, isTruncated := limitStringLength(body, 100)
+			if isTruncated {
+				truncatedBody += "... <truncated>"
+			}
+
+			slog.Error("Invalid response JSON in custom API widget", "url", req.httpRequest.URL.String(), "body", truncatedBody)
+			return nil, errors.New("invalid response JSON")
 		}
 
-		slog.Error("Invalid response JSON in custom API widget", "url", req.httpRequest.URL.String(), "body", truncatedBody)
-		return nil, errors.New("invalid response JSON")
+		return nil, fmt.Errorf("%d %s", resp.StatusCode, http.StatusText(resp.StatusCode))
+
 	}
 
-	data := &customAPIResponseData{
+	return &customAPIResponseData{
 		JSON:     decoratedGJSONResult{gjson.Parse(body)},
 		Response: resp,
-	}
-
-	return data, nil
+	}, nil
 }
 
-func fetchAndParseCustomAPI(
+func fetchAndRenderCustomAPIRequest(
 	primaryReq *CustomAPIRequest,
 	subReqs map[string]*CustomAPIRequest,
+	options customAPIOptions,
 	tmpl *template.Template,
 ) (template.HTML, error) {
 	var primaryData *customAPIResponseData
@@ -219,7 +287,7 @@ func fetchAndParseCustomAPI(
 
 	if len(subReqs) == 0 {
 		// If there are no subrequests, we can fetch the primary request in a much simpler way
-		primaryData, err = fetchCustomAPIRequest(context.Background(), primaryReq)
+		primaryData, err = fetchCustomAPIResponse(context.Background(), primaryReq)
 	} else {
 		// If there are subrequests, we need to fetch them concurrently
 		// and cancel all requests if any of them fail. There's probably
@@ -234,7 +302,7 @@ func fetchAndParseCustomAPI(
 		go func() {
 			defer wg.Done()
 			var localErr error
-			primaryData, localErr = fetchCustomAPIRequest(ctx, primaryReq)
+			primaryData, localErr = fetchCustomAPIResponse(ctx, primaryReq)
 			mu.Lock()
 			if localErr != nil && err == nil {
 				err = localErr
@@ -249,7 +317,7 @@ func fetchAndParseCustomAPI(
 				defer wg.Done()
 				var localErr error
 				var data *customAPIResponseData
-				data, localErr = fetchCustomAPIRequest(ctx, req)
+				data, localErr = fetchCustomAPIResponse(ctx, req)
 				mu.Lock()
 				if localErr == nil {
 					subData[key] = data
@@ -273,6 +341,7 @@ func fetchAndParseCustomAPI(
 	data := customAPITemplateData{
 		customAPIResponseData: primaryData,
 		subrequests:           subData,
+		Options:               options,
 	}
 
 	var templateBuffer bytes.Buffer
@@ -299,7 +368,7 @@ func gJsonResultArrayToDecoratedResultArray(results []gjson.Result) []decoratedG
 }
 
 func (r *decoratedGJSONResult) Exists(key string) bool {
-	return r.Get(key).Exists()
+	return r.Result.Get(key).Exists()
 }
 
 func (r *decoratedGJSONResult) Array(key string) []decoratedGJSONResult {
@@ -307,7 +376,7 @@ func (r *decoratedGJSONResult) Array(key string) []decoratedGJSONResult {
 		return gJsonResultArrayToDecoratedResultArray(r.Result.Array())
 	}
 
-	return gJsonResultArrayToDecoratedResultArray(r.Get(key).Array())
+	return gJsonResultArrayToDecoratedResultArray(r.Result.Get(key).Array())
 }
 
 func (r *decoratedGJSONResult) String(key string) string {
@@ -315,7 +384,7 @@ func (r *decoratedGJSONResult) String(key string) string {
 		return r.Result.String()
 	}
 
-	return r.Get(key).String()
+	return r.Result.Get(key).String()
 }
 
 func (r *decoratedGJSONResult) Int(key string) int {
@@ -323,7 +392,7 @@ func (r *decoratedGJSONResult) Int(key string) int {
 		return int(r.Result.Int())
 	}
 
-	return int(r.Get(key).Int())
+	return int(r.Result.Get(key).Int())
 }
 
 func (r *decoratedGJSONResult) Float(key string) float64 {
@@ -331,7 +400,7 @@ func (r *decoratedGJSONResult) Float(key string) float64 {
 		return r.Result.Float()
 	}
 
-	return r.Get(key).Float()
+	return r.Result.Get(key).Float()
 }
 
 func (r *decoratedGJSONResult) Bool(key string) bool {
@@ -339,7 +408,28 @@ func (r *decoratedGJSONResult) Bool(key string) bool {
 		return r.Result.Bool()
 	}
 
-	return r.Get(key).Bool()
+	return r.Result.Get(key).Bool()
+}
+
+func (r *decoratedGJSONResult) Get(key string) *decoratedGJSONResult {
+	return &decoratedGJSONResult{r.Result.Get(key)}
+}
+
+func customAPIDoMathOp[T int | float64](a, b T, op string) T {
+	switch op {
+	case "add":
+		return a + b
+	case "sub":
+		return a - b
+	case "mul":
+		return a * b
+	case "div":
+		if b == 0 {
+			return 0
+		}
+		return a / b
+	}
+	return 0
 }
 
 var customAPITemplateFuncs = func() template.FuncMap {
@@ -359,6 +449,31 @@ var customAPITemplateFuncs = func() template.FuncMap {
 		return regex
 	}
 
+	doMathOpWithAny := func(a, b any, op string) any {
+		switch at := a.(type) {
+		case int:
+			switch bt := b.(type) {
+			case int:
+				return customAPIDoMathOp(at, bt, op)
+			case float64:
+				return customAPIDoMathOp(float64(at), bt, op)
+			default:
+				return math.NaN()
+			}
+		case float64:
+			switch bt := b.(type) {
+			case int:
+				return customAPIDoMathOp(at, float64(bt), op)
+			case float64:
+				return customAPIDoMathOp(at, bt, op)
+			default:
+				return math.NaN()
+			}
+		default:
+			return math.NaN()
+		}
+	}
+
 	funcs := template.FuncMap{
 		"toFloat": func(a int) float64 {
 			return float64(a)
@@ -366,27 +481,59 @@ var customAPITemplateFuncs = func() template.FuncMap {
 		"toInt": func(a float64) int {
 			return int(a)
 		},
-		"add": func(a, b float64) float64 {
-			return a + b
+		"add": func(a, b any) any {
+			return doMathOpWithAny(a, b, "add")
 		},
-		"sub": func(a, b float64) float64 {
-			return a - b
+		"sub": func(a, b any) any {
+			return doMathOpWithAny(a, b, "sub")
 		},
-		"mul": func(a, b float64) float64 {
-			return a * b
+		"mul": func(a, b any) any {
+			return doMathOpWithAny(a, b, "mul")
 		},
-		"div": func(a, b float64) float64 {
+		"div": func(a, b any) any {
+			return doMathOpWithAny(a, b, "div")
+		},
+		"mod": func(a, b int) int {
 			if b == 0 {
-				return math.NaN()
+				return 0
+			}
+			return a % b
+		},
+		"now": func() time.Time {
+			return time.Now()
+		},
+		"offsetNow": func(offset string) time.Time {
+			d, err := time.ParseDuration(offset)
+			if err != nil {
+				return time.Now()
+			}
+			return time.Now().Add(d)
+		},
+		"duration": func(str string) time.Duration {
+			d, err := time.ParseDuration(str)
+			if err != nil {
+				return 0
 			}
 
-			return a / b
+			return d
 		},
-		"parseTime":      customAPIFuncParseTime,
+		"parseTime": func(layout, value string) time.Time {
+			return customAPIFuncParseTimeInLocation(layout, value, time.UTC)
+		},
+		"formatTime": customAPIFuncFormatTime,
+		"parseLocalTime": func(layout, value string) time.Time {
+			return customAPIFuncParseTimeInLocation(layout, value, time.Local)
+		},
 		"toRelativeTime": dynamicRelativeTimeAttrs,
 		"parseRelativeTime": func(layout, value string) template.HTMLAttr {
 			// Shorthand to do both of the above with a single function call
-			return dynamicRelativeTimeAttrs(customAPIFuncParseTime(layout, value))
+			return dynamicRelativeTimeAttrs(customAPIFuncParseTimeInLocation(layout, value, time.UTC))
+		},
+		"startOfDay": func(t time.Time) time.Time {
+			return time.Date(t.Year(), t.Month(), t.Day(), 0, 0, 0, 0, t.Location())
+		},
+		"endOfDay": func(t time.Time) time.Time {
+			return time.Date(t.Year(), t.Month(), t.Day(), 23, 59, 59, 0, t.Location())
 		},
 		// The reason we flip the parameter order is so that you can chain multiple calls together like this:
 		// {{ .JSON.String "foo" | trimPrefix "bar" | doSomethingElse }}
@@ -403,6 +550,13 @@ var customAPITemplateFuncs = func() template.FuncMap {
 		"replaceAll": func(old, new, s string) string {
 			return strings.ReplaceAll(s, old, new)
 		},
+		"replaceMatches": func(pattern, replacement, s string) string {
+			if s == "" {
+				return ""
+			}
+
+			return getCachedRegexp(pattern).ReplaceAllString(s, replacement)
+		},
 		"findMatch": func(pattern, s string) string {
 			if s == "" {
 				return ""
@@ -418,6 +572,7 @@ var customAPITemplateFuncs = func() template.FuncMap {
 			regex := getCachedRegexp(pattern)
 			return itemAtIndexOrDefault(regex.FindStringSubmatch(s), 1, "")
 		},
+		"percentChange": percentChange,
 		"sortByString": func(key, order string, results []decoratedGJSONResult) []decoratedGJSONResult {
 			sort.Slice(results, func(a, b int) bool {
 				if order == "asc" {
@@ -453,8 +608,8 @@ var customAPITemplateFuncs = func() template.FuncMap {
 		},
 		"sortByTime": func(key, layout, order string, results []decoratedGJSONResult) []decoratedGJSONResult {
 			sort.Slice(results, func(a, b int) bool {
-				timeA := customAPIFuncParseTime(layout, results[a].String(key))
-				timeB := customAPIFuncParseTime(layout, results[b].String(key))
+				timeA := customAPIFuncParseTimeInLocation(layout, results[a].String(key), time.UTC)
+				timeB := customAPIFuncParseTimeInLocation(layout, results[b].String(key), time.UTC)
 
 				if order == "asc" {
 					return timeA.Before(timeB)
@@ -464,6 +619,64 @@ var customAPITemplateFuncs = func() template.FuncMap {
 			})
 
 			return results
+		},
+		"concat": func(items ...string) string {
+			return strings.Join(items, "")
+		},
+		"unique": func(key string, results []decoratedGJSONResult) []decoratedGJSONResult {
+			seen := make(map[string]struct{})
+			out := make([]decoratedGJSONResult, 0, len(results))
+			for _, result := range results {
+				val := result.String(key)
+				if _, ok := seen[val]; !ok {
+					seen[val] = struct{}{}
+					out = append(out, result)
+				}
+			}
+			return out
+		},
+		"newRequest": func(url string) *CustomAPIRequest {
+			return &CustomAPIRequest{
+				URL: url,
+			}
+		},
+		"withHeader": func(key, value string, req *CustomAPIRequest) *CustomAPIRequest {
+			if req.Headers == nil {
+				req.Headers = make(map[string]string)
+			}
+			req.Headers[key] = value
+			return req
+		},
+		"withParameter": func(key, value string, req *CustomAPIRequest) *CustomAPIRequest {
+			if req.Parameters == nil {
+				req.Parameters = make(queryParametersField)
+			}
+			req.Parameters[key] = append(req.Parameters[key], value)
+			return req
+		},
+		"withStringBody": func(body string, req *CustomAPIRequest) *CustomAPIRequest {
+			req.Body = body
+			req.BodyType = "string"
+			return req
+		},
+		"getResponse": func(req *CustomAPIRequest) *customAPIResponseData {
+			err := req.initialize()
+			if err != nil {
+				panic(fmt.Sprintf("initializing request: %v", err))
+			}
+
+			data, err := fetchCustomAPIResponse(context.Background(), req)
+			if err != nil {
+				slog.Error("Could not fetch response within custom API template", "error", err)
+				return &customAPIResponseData{
+					JSON: decoratedGJSONResult{gjson.Result{}},
+					Response: &http.Response{
+						Status: err.Error(),
+					},
+				}
+			}
+
+			return data
 		},
 	}
 
@@ -476,7 +689,24 @@ var customAPITemplateFuncs = func() template.FuncMap {
 	return funcs
 }()
 
-func customAPIFuncParseTime(layout, value string) time.Time {
+func customAPIFuncFormatTime(layout string, t time.Time) string {
+	switch strings.ToLower(layout) {
+	case "unix":
+		return strconv.FormatInt(t.Unix(), 10)
+	case "rfc3339":
+		layout = time.RFC3339
+	case "rfc3339nano":
+		layout = time.RFC3339Nano
+	case "datetime":
+		layout = time.DateTime
+	case "dateonly":
+		layout = time.DateOnly
+	}
+
+	return t.Format(layout)
+}
+
+func customAPIFuncParseTimeInLocation(layout, value string, loc *time.Location) time.Time {
 	switch strings.ToLower(layout) {
 	case "unix":
 		asInt, err := strconv.ParseInt(value, 10, 64)
@@ -495,7 +725,7 @@ func customAPIFuncParseTime(layout, value string) time.Time {
 		layout = time.DateOnly
 	}
 
-	parsed, err := time.Parse(layout, value)
+	parsed, err := time.ParseInLocation(layout, value, loc)
 	if err != nil {
 		return time.Unix(0, 0)
 	}
