@@ -201,11 +201,12 @@ func Collect(req *SystemInfoRequest) (*SystemInfo, []error) {
 	// currently disabled on Windows because it requires elevated privilidges, otherwise
 	// keeps returning a single sensor with key "ACPI\\ThermalZone\\TZ00_0" which
 	// doesn't seem to be the CPU sensor or correspond to anything useful when
-	// compared against the temperatures Libre Hardware Monitor reports
-	// also disabled on openbsd because it's not implemented by go-psutil
-	if runtime.GOOS != "windows" && runtime.GOOS != "openbsd" {
+	// compared against the temperatures Libre Hardware Monitor reports.
+	// Also disabled on the bsd's because it's not implemented by go-psutil for them
+	if runtime.GOOS != "windows" && runtime.GOOS != "openbsd" && runtime.GOOS != "netbsd" && runtime.GOOS != "freebsd" {
 		sensorReadings, err := sensors.SensorsTemperatures()
-		if err == nil {
+		_, errIsWarning := err.(*sensors.Warnings)
+		if err == nil || errIsWarning {
 			if req.CPUTempSensor != "" {
 				for i := range sensorReadings {
 					if sensorReadings[i].SensorKey == req.CPUTempSensor {
@@ -227,35 +228,50 @@ func Collect(req *SystemInfoRequest) (*SystemInfo, []error) {
 		}
 	}
 
-	filesystems, err := disk.Partitions(false)
-	if err == nil {
-		for _, fs := range filesystems {
-			mpReq, ok := req.Mountpoints[fs.Mountpoint]
-			isHidden := req.HideMountpointsByDefault
-			if ok && mpReq.Hide != nil {
-				isHidden = *mpReq.Hide
-			}
-			if isHidden {
-				continue
-			}
-
-			usage, err := disk.Usage(fs.Mountpoint)
-			if err == nil {
-				mpInfo := MountpointInfo{
-					Path:        fs.Mountpoint,
-					Name:        mpReq.Name,
-					TotalMB:     usage.Total / 1024 / 1024,
-					UsedMB:      usage.Used / 1024 / 1024,
-					UsedPercent: uint8(math.Min(usage.UsedPercent, 100)),
-				}
-
-				info.Mountpoints = append(info.Mountpoints, mpInfo)
-			} else {
-				addErr(fmt.Errorf("getting filesystem usage for %s: %v", fs.Mountpoint, err))
-			}
+	addedMountpoints := map[string]struct{}{}
+	addMountpointInfo := func(requestedPath string, mpReq MointpointRequest) {
+		if _, exists := addedMountpoints[requestedPath]; exists {
+			return
 		}
-	} else {
-		addErr(fmt.Errorf("getting filesystems: %v", err))
+
+		isHidden := req.HideMountpointsByDefault
+		if mpReq.Hide != nil {
+			isHidden = *mpReq.Hide
+		}
+		if isHidden {
+			return
+		}
+
+		usage, err := disk.Usage(requestedPath)
+		if err == nil {
+			mpInfo := MountpointInfo{
+				Path:        requestedPath,
+				Name:        mpReq.Name,
+				TotalMB:     usage.Total / 1024 / 1024,
+				UsedMB:      usage.Used / 1024 / 1024,
+				UsedPercent: uint8(math.Min(usage.UsedPercent, 100)),
+			}
+
+			info.Mountpoints = append(info.Mountpoints, mpInfo)
+			addedMountpoints[requestedPath] = struct{}{}
+		} else {
+			addErr(fmt.Errorf("getting filesystem usage for %s: %v", requestedPath, err))
+		}
+	}
+
+	if !req.HideMountpointsByDefault {
+		filesystems, err := disk.Partitions(false)
+		if err == nil {
+			for _, fs := range filesystems {
+				addMountpointInfo(fs.Mountpoint, req.Mountpoints[fs.Mountpoint])
+			}
+		} else {
+			addErr(fmt.Errorf("getting filesystems: %v", err))
+		}
+	}
+
+	for mountpoint, mpReq := range req.Mountpoints {
+		addMountpointInfo(mountpoint, mpReq)
 	}
 
 	sort.Slice(info.Mountpoints, func(a, b int) bool {
